The music manager splits of from the mixer and uses an updated messaging system to mix the music into the rest
