+++ Sound/Sound/AudioEngine.cpp
@@ -227,7 +227,6 @@ inline Audio::Samples Audio::Engine::GetWorkerCurrentSampleCount()
 {
 	// To get the desired precision, the duration is casted
 	// to nanoseconds and then converted back to seconds
-	float temp = std::chrono::duration_cast<std::chrono::nanoseconds>
-		(std::chrono::steady_clock::now() - mWorkThreadStartTime).count() * 0.000000001f;
+	float temp = (std::chrono::steady_clock::now() - mWorkThreadStartTime).count() * 0.000000001f;
 	return Audio::ToSamples(temp);
 }
