+++ Sound/Sound/Sound.vcxproj
@@ -139,6 +139,7 @@
   <ItemGroup>
     <ClCompile Include="AudioBank.cpp" />
     <ClCompile Include="AudioFile.cpp" />
+    <ClCompile Include="AudioMusicManager.cpp" />
     <ClCompile Include="main.cpp">
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|x64'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Lib_d|x64'">true</ExcludedFromBuild>
